Split library explore and management views
